+++ CHANGELOG.md
@@ -6,6 +6,12 @@ Change are listed in reverse chronological order (newest to oldest).
 
 <span class="changelog">
 
+###### [ 1.0.20 ] - 2023/12/13
+
+  *  Added service "Reboot Device" - Reboots the operating system of the SoundTouch device.
+  *  Updated Remote Keypress service to include key_state argument.  This allows presets to be
+  *  Updated underlying `bosesoundtouchapi` package requirement to version 1.0.30.
+
 ###### [ 1.0.19 ] - 2023/12/04
 
   *  Updated underlying `bosesoundtouchapi` package requirement to version 1.0.22.